// src/app/modules/cpassapi/model/fornitore.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ListinoFornitore } from './listinoFornitore';


export interface Fornitore { 
    id?: string;
    codice?: string;
    naturaGiuridica?: string;
    ragioneSociale?: string;
    cognome?: string;
    nome?: string;
    codiceFiscale?: string;
    codiceFiscaleEstero?: string;
    partitaIva?: string;
    indirizzo?: string;
    numeroCivico?: string;
    cap?: string;
    comune?: string;
    provincia?: string;
    stato?: string;
    listinoFornitores?: Array<ListinoFornitore>;
    selected?: boolean;
}