// src/app/modules/cpassapi/model/testataOrdine.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Destinatario } from './destinatario';
import { Fornitore } from './fornitore';
import { Impegno } from './impegno';
import { Provvedimento } from './provvedimento';
import { Settore } from './settore';
import { Stato } from './stato';
import { StatoNso } from './statoNso';
import { TipoOrdine } from './tipoOrdine';
import { TipoProcedura } from './tipoProcedura';
import { Ufficio } from './ufficio';
import { Utente } from './utente';


export interface TestataOrdine { 
    id?: string;
    settore?: Settore;
    ufficio?: Ufficio;
    anno?: number;
    numero?: number;
    utenteCompilatore?: Utente;
    stato?: Stato;
    dataEmissione?: Date;
    dataConferma?: Date;
    dataAutorizzazione?: Date;
    dataScadenza?: Date;
    statoNso?: StatoNso;
    descrizione?: string;
    note?: string;
    tipoOrdine?: TipoOrdine;
    lottoAnno?: number;
    lottoNumero?: number;
    fornitore?: Fornitore;
    provvedimento?: Provvedimento;
    tipoProcedura?: TipoProcedura;
    numeroProcedura?: string;
    consegnaRiferimento?: string;
    consegnaDataDa?: Date;
    consegnaDataA?: Date;
    consegnaIndirizzo?: string;
    consegnaCap?: string;
    consegnaLocalita?: string;
    listImpegno?: Array<Impegno>;
    listDestinatario?: Array<Destinatario>;
    optlock?: string;
}